requirements.txt
boto3==1.13.22
botocore==1.16.22
docutils==0.15.2
flake8==3.8.2
jmespath==0.10.0
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
python-dateutil==2.8.1
PyYAML==5.3.1
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.9
